Add fatigue system
You can turn this feature off by pressing the M key. If the C and F value disappears from the window, it is disabled. To set the base fatigue, disable the fatigue function, adjust the level, and then press M again to enable the fatigue function. While the fatigue function is enabled, the V and B keys will decrease and increase the amount of fatigue each hit will add to the counter.