Bugfix: Key queue crash and "no window" crash
This update fixes the race condition key queue crash (where the keys to press changed as the keys are being pressed one by one, causing a variable corruption) and the "no window" crash, which happens when there are no windows named "Ragnarock " (with the space, as described by the game). This unshackles the program from the limits of pyautogui's synchronous method of handling the output, making it possible for Thrym to hit a set of very "tight" notes.