Thrym 2.3 - Add config function
This also addresses some of the wrong hue values causing the issue of notes not being detected due to a slight hue shift from the lights. Until a more elegant way of detecting false notes (especially the white fur of the crew clothing in the map "Midgard" and the area between the hull of the ship and the "water") can be made, small adjustments of the hue tolerance value will have to do.